requirements.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.5.18.1
cloudinary==1.29.0
confusable-homoglyphs==3.2.0
coverage==6.4.1
dj-database-url==0.5.0
Django==4.0.5
django-appconf==1.0.5
django-debug-toolbar==3.4.0
django-imagekit==4.1.0
django-registration==3.3
django-sass==1.1.0
gunicorn==20.1.0
libsass==0.21.0
pilkit==2.0
Pillow==9.1.1
psycopg2==2.9.3
pycodestyle==2.8.0
python-decouple==3.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
